03._Assignments/server_1/pyproject.toml
[project]
name = "server-1"
version = "0.1.0"
description = ""
authors = [
    {name = "DizzyMoon",email = "mikkel.simons@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "xmltodict (>=0.14.2,<0.15.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "server_1", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
